better profile modal ui, single modal component, paid_dues default value in member controller
